+++ DuckDB.NET.Data/Internal/Writer/IntervalVectorDataWriter.cs
@@ -6,5 +6,5 @@ namespace DuckDB.NET.Data.Internal.Writer;
 
 internal sealed unsafe class IntervalVectorDataWriter(IntPtr vector, void* vectorData, DuckDBType columnType) : VectorDataWriterBase(vector, vectorData, columnType)
 {
-    internal override bool AppendTimeSpan(TimeSpan value, ulong rowIndex) => AppendValueInternal((DuckDBInterval)value, rowIndex);
+    internal override bool AppendTimeSpan(TimeSpan value, int rowIndex) => AppendValueInternal((DuckDBInterval)value, rowIndex);
 }
