+++ DuckDB.NET.Data/Internal/Writer/StringVectorDataWriter.cs
@@ -5,16 +5,16 @@ namespace DuckDB.NET.Data.Internal.Writer;
 
 internal sealed unsafe class StringVectorDataWriter(IntPtr vector, void* vectorData, DuckDBType columnType) : VectorDataWriterBase(vector, vectorData, columnType)
 {
-    internal override bool AppendString(string value, ulong rowIndex)
+    internal override bool AppendString(string value, int rowIndex)
     {
         using var unmanagedString = value.ToUnmanagedString();
-        NativeMethods.Vectors.DuckDBVectorAssignStringElement(base.Vector, rowIndex, unmanagedString);
+        NativeMethods.Vectors.DuckDBVectorAssignStringElement(base.Vector, (ulong)rowIndex, unmanagedString);
         return true;
     }
 
-    internal override bool AppendBlob(byte* value, int length, ulong rowIndex)
+    internal override bool AppendBlob(byte* value, int length, int rowIndex)
     {
-        NativeMethods.Vectors.DuckDBVectorAssignStringElementLength(Vector, rowIndex, value, length);
+        NativeMethods.Vectors.DuckDBVectorAssignStringElementLength(Vector, (ulong)rowIndex, value, length);
         return true;
     }
 }
